OPENCL FLOAT ONLY, REMOVED ALL DOUBLE

+++ include/caffe/util/OpenCL/OpenCLSupport.hpp
@@ -108,7 +108,7 @@ namespace caffe {
 namespace OpenCL {
 
 	int64_t getTickCount();
-	double getTickFrequency();
+	float getTickFrequency();
 
 //	bool init();
