+++ src/caffe/test/test_convolution_layer.cpp
@@ -96,10 +96,10 @@ template void caffe_conv(const Blob<float>* in,
     ConvolutionParameter* conv_param,
     const vector<shared_ptr<Blob<float> > >& weights,
     Blob<float>* out);
-template void caffe_conv(const Blob<double>* in,
+template void caffe_conv(const Blob<float>* in,
     ConvolutionParameter* conv_param,
-    const vector<shared_ptr<Blob<double> > >& weights,
-    Blob<double>* out);
+    const vector<shared_ptr<Blob<float> > >& weights,
+    Blob<float>* out);
 
 template <typename TypeParam>
 class ConvolutionLayerTest : public MultiDeviceTest<TypeParam> {
